+++ example/lib/injector.dart
@@ -1,4 +1,3 @@
-import 'package:example/service.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
@@ -7,15 +6,7 @@ import 'injector.iconfig.dart';
 final getIt = GetIt.instance;
 
 @injectableInit
-void configure() async => $initGetIt(getIt);
+Future<void> configure() => $initGetIt(getIt);
 
-@RegisterModule
-abstract class Module {
-  @Bind(ServiceImpl2)
-  Service get serviceImpl2;
-
-  @Bind(ServiceImpl2)
-  Service get serviceImpl22;
-
-  Future<Service> get prefs => Future.value();
-}
+@registerModule
+abstract class RegisterModule {}