+++ example/lib/main.dart
@@ -1,11 +1,7 @@
+import 'package:example/injector.dart';
+import 'package:flutter/material.dart';
+
 void main() {
-	// print(services);
-	// services.sort((a, b) {
-	//   print(
-	//       "A ${a.runtimeType} : ${a.deps.map((d) => d.runtimeType)}   B ${b.runtimeType}");
-	//   return a.deps.contains(b) ? -1 : 1;
-	// });
-	// services.forEach((s) => print(s));
-	// configure();
-	// runApp(MyApp('s'));
+  configure();
+  runApp(MaterialApp());
 }