+++ injectable_generator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ChangeLog
 
+## [0.2.0] Breaking Changes!
+
+- You now need to pass in your getIt instance to @initGetIt() func
+- Rename @Bind to @RegisterAs to avoid confusion because,
+  now we're annotating the implementation not the abstract.
+- Add @factoryMethod annotation to mark named factories and static create methods.
+- Add @registerModule annotation to support registering third party types.
+- Fix eager singletons are registered before their dependencies.
+
 ## [0.1.0]
 
 - initial release